clean up imports
This commit cleans up imports. It removes imports
from datalad.cmd that were just used by other modules.

The commit also removes a cyclic dependency that appears
when CapturedException is imported at top-level in
datalad.runner.exceptions.py if datalad.utils.ensure_unicode
is imported in global scope. The import of ensure_unicode
has been moved to global scope to allow CommandError
formatting even when Python is shuting down and it cannot
longer be imported in the functions that use it. For
the same reason config variable reading in methods
of BatchedCommmand has been moved to global scope.