fix description of runner and protocols
This commit fixes the description of
datalad.runner.protocol.WitlessProtocol,
of datalad.runner.noasyncrunner.ThreadedRunner,
and of datalad.runner.runner.WitlessRunner,
to better describe the semantics of the
stdin argument.

The type annotations of the stdin arguments
of datalad.runner.runner.WitlessRunner and
of datalad.runner.noasyncrunner.ThreadedRunner
are fixed to properly represent the expected
types.